+++ packages/auth/ios/RNFBAuth/RNFBAuthModule.m
@@ -1126,6 +1126,10 @@ RCT_EXPORT_METHOD(useEmulator
     credential = [FIROAuthProvider credentialWithProviderID:@"oauth"
                                                     IDToken:authToken
                                                 accessToken:authTokenSecret];
+  } else if ([provider hasPrefix:@"oidc."]) {
+    credential = [FIROAuthProvider credentialWithProviderID:provider
+                                                    IDToken:authToken
+                                                   rawNonce:nil];
   } else {
     DLog(@"Provider not yet handled: %@", provider);
   }