+++ packages/firestore/e2e/Query/where.e2e.js
@@ -88,6 +88,10 @@ describe('firestore().collection().where()', function () {
     firebase.firestore().collection(COLLECTION).where('foo.bar', '==', null);
   });
 
+  it('allows null to be used with not equal operator', function () {
+    firebase.firestore().collection(COLLECTION).where('foo.bar', '!=', null);
+  });
+
   it('throws if multiple inequalities on different paths is provided', function () {
     try {
       firebase.firestore().collection(COLLECTION).where('foo.bar', '>', 123).where('bar', '>', 123);
@@ -292,6 +296,42 @@ describe('firestore().collection().where()', function () {
     snapshot.forEach(s => {
       s.data().foo.should.be.belowOrEqual(search);
     });
+  });
+
+  it('returns when combining greater than and lesser than on the same nested field', async function () {
+    const colRef = firebase.firestore().collection(`${COLLECTION}/filter/greaterandless`);
+
+    await Promise.all([
+      colRef.doc('doc1').set({ foo: { bar: 1 } }),
+      colRef.doc('doc2').set({ foo: { bar: 2 } }),
+      colRef.doc('doc3').set({ foo: { bar: 3 } }),
+    ]);
+
+    const snapshot = await colRef
+      .where('foo.bar', '>', 1)
+      .where('foo.bar', '<', 3)
+      .orderBy('foo.bar')
+      .get();
+
+    snapshot.size.should.eql(1);
+  });
+
+  it('returns when combining greater than and lesser than on the same nested field using FieldPath', async function () {
+    const colRef = firebase.firestore().collection(`${COLLECTION}/filter/greaterandless`);
+
+    await Promise.all([
+      colRef.doc('doc1').set({ foo: { bar: 1 } }),
+      colRef.doc('doc2').set({ foo: { bar: 2 } }),
+      colRef.doc('doc3').set({ foo: { bar: 3 } }),
+    ]);
+
+    const snapshot = await colRef
+      .where(new firebase.firestore.FieldPath('foo', 'bar'), '>', 1)
+      .where(new firebase.firestore.FieldPath('foo', 'bar'), '<', 3)
+      .orderBy(new firebase.firestore.FieldPath('foo', 'bar'))
+      .get();
+
+    snapshot.size.should.eql(1);
   });
 
   it('returns with where array-contains filter', async function () {